requirements.txt
llama-index
openai
tiktoken
python-dotenv
llama-index-llms-ollama
flask